Create ScumFileManager
Given information about a game and save, ScumFileManager is responsible for determining exactly where backups will be stored, as well as responsible for the actual heavy lifting of creating the backup files and restoring them when requested. The relieves the burden of actually managing files, from other classes. Instead they can focus on determining which files should be backed up or restored, and when.